Models fixed
Models can now be changed between runs.

Really this just required a little edit to how model selection was done, and a couple of checks to make sure model-selection is re-done if the model-list differs from what you want it to be.

Basically it just separates out the models you've analysed for a subset (subset.results{}) from those you're interested in right now.

Brett can you check this out - I remember you said there was something fundamental that would have to change when we did this, but I didn't come across it. Maybe I've buggered it up. Hence the branch.

Still it seems to work for now, either when I add models or take them away from what happened in previous runs.

R