Changed [partitions] to [data_blocks] throughout
This is a bit tedious, but it makes everything clearer, and means that e
verything in the program now matches what we write in the paper (current
ly under review).

Note that this is only true from the user's perspecti
ve. Within the program itself, we still call a 'partition' a set of site
s. However, the user will never see this, they will see 'data block', bo
th in the config file and in the PF output.

R