Fixed AICc calculation and added warnings
Our AICc calcultion was still a bit off for very small subsets. It is
now fixed. I have also added warnings to the user to check carefully
when they use the AICc with small subsets, because we use a fairly
simple (but sensible) fix to get around the problem without crashing
each time.