added tree_size option to RAxML parser
Note that this doesn't yet work all the time. This is because the
output files for RAxML when used with -f B don't contain tree_size
information, and there's no tree file to get the information from
either. Right now there's nothing we can do about this, but hopefully
Alexis will help out. :)