added neighbour.py to find neighbouring schemes
this is very crudely written, but it works. It takes a scheme, and
returns the scheme which has the two most similar subsets clustered.
The useful method is get_nearest_neighbour_scheme, we pass in a scheme,
the name for the new scheme, a cfg object, and some weights. The
weights allow us to put different importance on different parameters,
e.g.

weights = {"rate": 1, "freqs": 2, "model": 10}

This will require some (a lot?) empirical testing, but the main thing
will be to set sensible defaults, then add a commandline option to
change them if the user wants.