added two stopping conditions for greediest algorithm
--greediest-schemes
this sets the number of major improvements (>10 units of BIC/AIC/AICc)
we will wait for before progressing

--greediest-percent
if we have analysed at least this many percent of the possible schemes
in a given step of the greediest algorithm, we just take the best
answer we can find and move on. However, if we didn't find ANY
improvement, we keep looking at take the next one we find.