Fixed long line '\' breaks in pyparsing
This issue was identified by Johan Nylander - it seems to stop PF
working on Xubuntu. Better just to get rid of all of the '\' line
continuations, and have some long lines in pyparsing. Not sure why this
was limited to Xubuntu - PF works fine on some other linux systems.
Hmmm.