streamlining of greediest algorithm
now there's only one implemented stopping condition - percent. So, you
just pass the --greediest_percent argument at the commandline. The
clustering paper will allow us to set sensible defaults for this, but
in the absence of any other information I think 10% is a good starting
point. For massive massive datasets it will have to be set lower, as
pragmatism demands!