clunky fix for progress reporting
the real problem seems to be that subsets don't remember that they're
DONE. So, if I recreate an existing subset, the factory appears to load
it up, but that subset doesn't know that it was already done.

This workaround is bad, because it keeps a set of subsets, which takes
too much memory.

We need a better fix. Not sure what.