changed search methods to strict_clustering and relaxed_clustering
Right now some tests fail, for two different reasons.

1. It seems that having underscores in the .cfg file isn't allowed.
This is easy to fix (we can rename them again), but need to check if
there's a simple way to allow for underscores first.

2. We have a bug in our version of RAxML, in some cases we get this
error:

Assertion failed: (inputLikelihood == tr->likelihood), function modOpt,
file optimizeModel.c, line 3233.

I chatted with Alexis about this before. First we need to make 100%
sure we have the very latest version of RAxML. If we do, then we need
to get in touch with Alexis to see what the issue might be.

R