fix a file deletion bug on windows
The bug occurred randomly and rarely during threading. It was pointed
out by Ryan Kepler, and also noticed by others at a NESCent workshop.

The fix is to try deleting files, and just not worry about errors. The
absolute worst that happens here is that we end up keeping some files
that we wanted to delete. This is not bad. However, it seems that the
bug is actually in Python, and that even though an error is sometimes
thrown when deleting files in threads, the file gets deleted anyway.