make reporter write out simple site lists
there was nothing wrong with brett's solution, but in the end it makes
basically no difference to the complexity of the parititoin definitions
that are output, and it makes post-processing those definitions to look
at plots much harder. So I canned it and now we have a simple solution.

Before:

p1 = 1, 4, 7-10, 12

After

p1 = 1, 4, 7, 8, 9, 10, 12