don't ignore errors
This is (i think) something that Paul put in to get the kmeans approach
to work. But it's a bad idea - it overlooks all errors in phylogeny
programs, and continues as if they didn't occur. This causes all sorts
of issues, and we'll have to find a better way to do this. For now, we
revert to the sensible approach.