count subsets by step, not overall
the counting is really misleading for many algorithms - the rcluster
algorithm tends to finish after ~1-2% of all possible schemes. This is
because the overlal count is based on worse-case scenarios, which are
almost never realised. In the end it's almost impossible to predict the
% done from the subsets.

Much better to count it from the number of steps done, at least for the
rcluster algorithm.