update download stats and plots
also added a new plot which, just for fun, calculates the number of
downloads per citation. This obviously isn't very accurate, since
downloads in a given month do not lead to citations in the same month,
but it's an interesting idea.

Perhaps give some indication of the rate of attrition (and/or re-use)
of users. I.e. how many download it then give up and never use it for a
publication.