include Likelihood Ratio Test function
this is an idea suggested by Olivier Gascuel. The issue is that the
number of partitioning schemes with lots of subsets is practically
infinite. The upshot is that we can almost always find a schemes with
more subsets that are very slightly better than the current scheme.
This causes the problem that we over-split datasets.

We can see this on simulated data, as we point out in the BMC paper.
One way to avoid this is to perform a likelihood ratio test each time
we split a subset, and only accept the split if the LRT is significant.