removed LRT
because I realised that the models we are comparing are not necessarily
nested. The point is that although they are nested in terms of the
partitions, it’s possible for the model with fewer partitions to have
more complex substitution models. So there is no guarantee of
nestedness, which means we can’t use LRTs. An alternative might be to
use a threshold for delta AICc/BIC.