fix bug in get_best_pair
this bug occurs because (very rarely) we incorrectly looked up the
address in the matrix of information theory score improvements.

The result was that we would sometimes choose the same subset twice,
and this had repurcussions which caused PF to crash when cleaning up
subsets at the end of a step of the rcluster algorithm.

I added a log of debug output to this function.