changed subset chunk limit to 1000
This is a little trick to reduce memory usage in greedy analyses. If
the list of subsets to analyse gets big, then it uses a lot of memory.
The fix we use for this is to monitor the list, and clear it out
periodically when it crosses a threshold.

The previous threshold was 10K, but in testing I noticed that it could
take a while to build 10K subsets. This isn’t really an issue, except
that it looks like PF2 isn’t doing anything much.

Reducing it to 1K means that we spend less time waiting. It’s probably
marginally slower (by utterly unimportant margins), but it will likely
be more memory efficient.