set default rcluster-percent back to 10
Here’s why. In tests I noticed that having rcluster-max at 1000, and
rcluster-percent at 100 meant that the algorithm was actually pretty
slow in the later stages. This is because it ends up examining a lot
more of the bigger subsets we encounter at the end of a run.

By setting it back to 10 percent, we will gain speed and (I think) lose
no meaningful performance.