fixed bug in rcluster
This is really fixed now. The issue was that the previous bug fix
wasn’t bulletproof. It left the door open for a second bug, in which a
single subset had an identical improvement score with >1 other subset.

The new fix addresses this bug, as well as making sure that the
original bug is fixed.