implemented --min-subset-size option
The default is 100.

The idea is that very small subsets are not biologically (or
algorithmically) sensible. The kmeans algorithm left to its own devices
likes to make such subsets, which is not a very good idea, and causes
various downstream issues.

So, we now have a minimum subset size. Which is implemented as follows:

1. If a new subsets is < minimum, we don’t try and split it any further
2. If at the end of the run we have subsets < minimum, we call them
fabricated and merge them back to the most similar remaining subset.