only split subsets if there are no state probs in the daughters
similar to the min subset length, we only keep the split if neither of
the daughter subsets have issues that the user doesn’t want.

It seems better to do this, than to keep those subsets and try and
merge them back in at the end, because I suspect that the merging is
not very accurate.