update rcluster to check for problematic subsets
based on the —all-states and —min-subset-size command line parameters,
users can specify whether PartitionFinder should return only those
subsets with all states, or above a size threshold.

This is already implemented for kmeans, but this commit implements it
for rcluster too.

The behaviour is as follows.

We start by analysing the starting scheme (note - this means that PF
has to be able to analyse all of the user-specified subsets).

We the check through all the subsets in the starting scheme (smallest
subsets first). As soon as we find a subset that fails either of the
above tests (as specified by the user), we merge it with its nearest
neighbour subset. In this case, the nearest neighbour subset is defined
as simply the other subest that is nearest based on the subset
similarity metrics the user has specified (i.e. by the —weights). Thus,
the behaviour is consistent with whatever the user has decided is a
sensible measure of subset similarity.

Once the first such subset has been fixed, we start again from scratch
- analyse the new scheme, get the list of subsets, sort them smallest
to largest, look for problems.

Once we find no more problems, we continue.

If we merge all subsets at this step, we report that and tell the user.