--min-subset-size defaults
The default for kmeans remains 100
The default for rcluster is to ignore this option. Thus, for rcluster
if you want to use —min-subset size you have to specify it explicitly.
This is so that the default behaviour of the rcluster algorithm remains
faithful to the description in the paper, to save confusion.