change fcluster to rclusterf, and remove duplicate code
This merges the rcluster and rclusterf (previously fcluster) analyses
into a single class, to make it clear (both in the code and to users)
that rclusterf really is just a variant of rcluster.

Of course, one could have a parameter for how many subset to join at
each step of the analysis, but this seems like it’s going too far.