add functions to raxml.py to calculate ML tree with pthreads
we use pthreads because in this case it makes a lot more sense.

Currently the default behaviour is to use GTR+G or LG+G, with linked
branch lengths and a separate model for each partition. Partitions are
just defined by the users starting scheme. This does two things:

1. Following from david and my paper, in which we show that just using
all the partitions is pretty much never wrong in an interesting way,
this approach should mean that the starting tree is almost always the
best possible tree.

2. Following from Paul and my work (and others) in which we show that
in some rare cases with model selection, one can observe a starting
tree bias, this should do a LOT better at avoiding that in almost all
cases.