update binaries for PhyML and RAxML
updated to PhyML v3.1 and RAxML latest commit (as of today).

Wrote a README to explain how to get correct versions on various
systems.

Updated phyml parsers to reflect the fact that files now have ‘.txt’
appended.

included binaries for Mac, Windows, Linux.

All but 5 tests pass. The 5 failures are just because of some small
diffs in likelihood with the new version of PhyML.