new krmeans algorithm
the previous version was naive. I reassigned invariant sites at every
step, which just got the algorithm stuck early on. This version waits
until the end of the kmeans algorithm to reassign sites, which is a
much better idea. It appears to work well (in terms of AICc scores) on
empirical datasets.