change default rcluster-max
The old default was 1000, but this was too small for some datasets. The
new default is the largest out of 1000 and 20* the number of data
blocks.

This means that the new rcluster is still linear time (compared to
exponential time for the old one), but that it makes a better balance
between being quick and getting good answers.