implement IQtree output format
following the description on the IQtree website. At this point it’s not
clear which models might or might not be available in IQtree, so I do
not yet have an IQtree model list. Instead, I just output the best
model for each subset, and rely on users to specify appropriate models.