Disable k-means for all but morphology
#Why?

A paper came out yesterday
(http://www.sciencedirect.com/science/article/pii/S1055790316302780)
that raises some serious concerns about the k-means algorithm,
suggesting that it might lead to bad inferences on empirical datasets.
I had spoken to the authors of the paper when they were revising it,
but wasn’t aware until yesterday of the details of the problems they’d
uncovered. Given how odd the inferences from k-means look, we decided
to disable the method for all but morphological analyses (see below).

# But there was a warning before, why disable it now?

Our previous concerns came from our own realisation about one aspect of
the method (that it lumps together all invariant sites) and some
concerns raised by folks in Brian Moore’s lab this year. Specifically,
we put in the warning when we learned that some simulated datasets that
were analysed with k-means partitioning schemes led to bad inferences.

I was hopeful that these simulations would be corner cases, and/or that
one aspect of the simulations where k-means was misleading (that you
got implausibly long trees) would mean that it would be trivial to
diagnose cases in which there were issues. In addition, we had tried
the method on lots of empirical datasets, and never seen any issues.
Indeed, on at least one dataset the k-means tree seemed much more
reasonable than trees we were getting from other methods. (I note that
Brian Moore and co were less optimistic, and suggested from the start
that we should consider disabling the method.)

The empirical results in the recent paper suggest otherwise, and
suggest that the best we can say of k-means for now is: ‘you should try
other methods too, and if the methods disagree, we’d suggest ignoring
the k-means tree'. On this basis, there seems little point keeping
k-means as an available method: if you can't trust the reuslts, why
bother.

# I liked/used it, what should I do?

Use standard methods, e.g. partitioning by codon position and locus,
instead. Even better (if your dataset is small enough) use the
automatic partitioning solutions in BEAST2 and/or MrBayes (google
AutoParts).

If you have used k-means to make an inference, it would be worthwhile
to check that the inference is robust when you use a standard
partitioning scheme too.

# What’s the problem?

We don’t know for sure, but it’s likely to be related to the fact that
k-means separates out all invariant sites into a single subset. I
presented on this at SMBE in July this year, but this has a couple of
downstream effects. First, it makes AIC/AICc/BIC scores look really
great, because when you have all the invariant sites together, you can
estimate a rate of zero and get likelihoods of 1 for all of those
sites. That’s a bit silly, and something I wish we’d realised earlier.
Second, and more seriously for inference, putting all the invariant
sites into one subset means that the other subsets have NO invariant
sites. If you then analyse these without a model that accounts for this
(e.g. with some kind of ascertainment bias) this is likely to mess with
estimate of rates, branch lengths, and topologies. It’s not totally
obvious yet how common the problem is, but now we’ve seen it in
simulated and empirical datasets, it seems wise to can the method until
we completely understand the problem and can fix it.

# Are you going to fix it?

We're working on it, but it will take a while. Apart from anything
else, we are going to be exceptionally cautious in proposing more new
methods related to this one.

# But why is it still available for morphology?

We’ve kept it in there for morphological datasets as an experimental
method, and provide lots of warnings when you run the code and in the
output that it’s experimental, untested etc. We did this because
morphological datasets are different: they tend to have no invariant
sites, and people tend to use models that correct for ascertainment
bias. Because of that, it seems worthwhile to leave it in. We are
working on testing it as exhaustively as possible for these datasets.

# I want to use it anyway

If you want to use it for empirical inferences, just don’t. But if you
want to use it to try and figure out why it doesn’t work, and how you
might improve it, then all you need to do is edit out the line that
raises the error.

# I have questions…

Post on the google group or raise an issue on GitHub.