Update chinese translation of manual
It came to my attention that the translation included links to a paid
discussion forum. There is nothing particularly wrong with charging
people for a discussion forum, in the sense that it’s permissible under
the software license. But we have a very good forum, and there is no
need for people to pay for advice on partition finder, which can be
obtained for free.